add custom order select page
