+++ lib/constants.dart
@@ -1,1 +1,2 @@
 const defaultScene = "よく使うフレーズ";
+const customScene = "その他の場面";
