+++ lib/ui/page/order/select/custom/page.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+class CustomOrderSelectPage extends ConsumerWidget {
+  const CustomOrderSelectPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return const Column(children: [Text("カスタム")]);
+  }
+}
